+++ legacy-data-access-services/src/main/java/gov/ca/cwds/cms/data/access/dto/OtherAdultInHomeEntityAwareDTO.java
@@ -1,9 +1,7 @@
 package gov.ca.cwds.cms.data.access.dto;
 
 import gov.ca.cwds.data.legacy.cms.entity.OtherAdultsInPlacementHome;
-import gov.ca.cwds.data.legacy.cms.entity.OtherPeopleScpRelationship;
 import gov.ca.cwds.data.legacy.cms.entity.OutOfStateCheck;
-
 import java.util.ArrayList;
 import java.util.List;
 
@@ -12,30 +10,16 @@ import java.util.List;
  */
 
 public class OtherAdultInHomeEntityAwareDTO extends
-    BaseEntityAwareDTO<OtherAdultsInPlacementHome> {
+  RelationshipsAwareDTO<OtherAdultsInPlacementHome> {
 
-  private List<OtherPeopleScpRelationship> relationships = new ArrayList<>();
   private List<OutOfStateCheck> outOfStateChecks = new ArrayList<>();
-
-  public List<OtherPeopleScpRelationship> getRelationships() {
-    return relationships;
-  }
-
-  public void setRelationships(
-      List<OtherPeopleScpRelationship> relationships) {
-    this.relationships = relationships;
-  }
-
-  public void addRelationship(OtherPeopleScpRelationship relationship) {
-    relationships.add(relationship);
-  }
 
   public List<OutOfStateCheck> getOutOfStateChecks() {
     return outOfStateChecks;
   }
 
   public void setOutOfStateChecks(
-      List<OutOfStateCheck> outOfStateChecks) {
+    List<OutOfStateCheck> outOfStateChecks) {
     this.outOfStateChecks = outOfStateChecks;
   }
